WildTangz: How the art was born!
This commit represents the script used to make the final collection
(with some minor hand tweaks) of WildTangz (policy ID
33568ad11f93b3e79ae8dee5ad928ded72adcea719e92108caf1521b on the Cardano
blockchain).

The overview is that it uses an iterative goal-seek to continually
optimize for maximum uniqueness while targeting specific attribute
rarities.  Future versions of this library will be made more useful and
generic.

[ Documentation: None yet ]
[ Testing: None ]