feat(regenerate-directory): Add ability to recursively regenerate images
This is very useful when upgrading a collection as we will have all the
v1 traits, but new artwork.  So, this change allows us to regenerate a
directory from the command line.

NOTE: Minor change from webp -> png since Cardano IPFS is not
constrained by size limits a la Bitcoin.

[ Testing: python3 src/py/compose.py regenerate-directory --... ]