feat(Migration): Add a script to migrate from paid IPFS to free storage
Because IPFS represents files as a directed acyclic graph (DAG) of
Content Archive (CAR) files, we can't just re-upload directly to free
storage like NFT.Storage (the chunking is likely different).

The algorithm here is:
1. Retrieve assets on chain or from a specified metadata directory
2. Directly export the CAR from the local IPFS daemon for each NFT
2. Re-upload to NFT.Storage as a CAR file, instead of a raw image

NOTE: This requires `ipfs daemon` to be running so that the user can
retrieve CAR files.

[ Testing: Manual run for images ]